test: split huge cases file to multiple files
